fix: move utils integration test
