+++ dataworkspace/dataworkspace/tests/core/test_utils.py
@@ -1,12 +1,15 @@
 import datetime
 import re
+import time
 
+import mock
 import psycopg2
 import pytest
 from django.conf import settings
 from django.db import connections
 from django.test import override_settings
 
+from dataworkspace.apps.applications.utils import delete_unused_datasets_users
 from dataworkspace.apps.core.models import DatabaseUser
 from dataworkspace.apps.core.utils import (
     database_dsn,
@@ -152,3 +155,61 @@ class TestNewPrivateDatabaseCredentials:
             results = cursor.fetchall()
             assert 'pgaudit.log=ALL' in results[0][1]
             assert 'pgaudit.log_catalog=off' in results[0][1]
+
+
+class TestDeleteUnusedDatasetsUsers:
+    @pytest.mark.django_db(transaction=True)
+    def test_deletes_expired_and_unused_users(self):
+        ensure_databases_configured().handle()
+
+        user = UserFactory(email='test@foo.bar')
+        st = SourceTableFactory(
+            dataset=MasterDataSetFactory.create(
+                user_access_type='REQUIRES_AUTHENTICATION'
+            )
+        )
+
+        source_tables = source_tables_for_user(user)
+        db_role_schema_suffix = db_role_schema_suffix_for_user(user)
+        user_creds_to_drop = new_private_database_credentials(
+            db_role_schema_suffix,
+            source_tables,
+            postgres_user(user.email),
+            user,
+            valid_for=datetime.timedelta(days=31),
+        )
+        qs_creds_to_drop = new_private_database_credentials(
+            db_role_schema_suffix,
+            source_tables,
+            postgres_user(user.email, suffix='qs'),
+            user,
+            valid_for=datetime.timedelta(seconds=0),
+        )
+        qs_creds_to_keep = new_private_database_credentials(
+            db_role_schema_suffix,
+            source_tables,
+            postgres_user(user.email, suffix='qs'),
+            user,
+            valid_for=datetime.timedelta(minutes=1),
+        )
+
+        connections[st.database.memorable_name].cursor().execute('COMMIT')
+
+        # Make sure that `qs_creds_to_drop` has definitely expired
+        time.sleep(1)
+
+        with mock.patch('dataworkspace.apps.applications.utils.gevent.sleep'):
+            delete_unused_datasets_users()
+
+        with connections[st.database.memorable_name].cursor() as cursor:
+            cursor.execute(
+                "SELECT usename FROM pg_catalog.pg_user WHERE usename IN %s",
+                [
+                    (
+                        user_creds_to_drop[0]['db_user'],
+                        qs_creds_to_drop[0]['db_user'],
+                        qs_creds_to_keep[0]['db_user'],
+                    )
+                ],
+            )
+            assert cursor.fetchall() == [(qs_creds_to_keep[0]['db_user'],)]
